% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat.R
\name{ipcc_cellchat1}
\alias{ipcc_cellchat1}
\title{Using cellchat to analysis.1}
\usage{
ipcc_cellchat1(input, species)
}
\arguments{
\item{input}{your scRNA}

\item{species}{mouse or human}
}
\value{
pathwayname
}
\description{
Using cellchat to analysis.1
}
\examples{
ipcc_cellchat1(input = scRNA,specis="mouse")
}
