% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellchat.R
\name{ipcc_cellchat2}
\alias{ipcc_cellchat2}
\title{Using cellchat to analysis.2}
\usage{
ipcc_cellchat2(input, species, path)
}
\arguments{
\item{input}{your scRNA}

\item{species}{mouse or human}

\item{path}{input pathway name from console}
}
\value{
figures
}
\description{
Using cellchat to analysis.2
}
\examples{
ipcc_cellchat2(input=scRNA,specis="mouese",path="GALECTIN")
}
