% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdb.R
\name{ipcc_cpdb}
\alias{ipcc_cpdb}
\title{ipcc_cpdb,using cellphonedb to analysis}
\usage{
ipcc_cpdb(input, species)
}
\arguments{
\item{input}{your scRNA}

\item{species}{mouse or human}
}
\value{
figures and txt analyzed by cellphonedb
}
\description{
ipcc_cpdb,using cellphonedb to analysis
}
\examples{
ipcc_cpdb(input= scRNA,species = "mouse")
}
