% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenet.R
\name{ipcc_nichenet}
\alias{ipcc_nichenet}
\title{Using nichenet to analysis}
\usage{
ipcc_nichenet(input, species, control, case)
}
\arguments{
\item{input}{you scRNA}

\item{species}{mouse or human}

\item{control}{the name of your control group}

\item{case}{the name of your case group}
}
\value{
figures and txt
}
\description{
Using nichenet to analysis
}
\examples{
ipcc_nichenet (input=scRNA,species="mouse",control="SS",case="LCMV")
}
